client: move multiplexer to separate file
